refactor: put on package "Payment" all of context payment
